stellar coin module
