+++ libs/coin-modules/coin-stellar/.unimportedrc.json
@@ -1,0 +1,16 @@
+{
+  "entry": [
+    "src/api/index.ts",
+    "src/bridge/js.ts",
+    "src/cli-transaction.ts",
+    "src/deviceTransactionConfig.ts",
+    "src/errors.ts",
+    "src/hw-getAddress.ts",
+    "src/serialization.ts",
+    "src/specs.ts",
+    "src/transaction.ts"
+  ],
+  "ignorePatterns": ["**/node_modules/**", "**/*.fixture.ts", "**/*.mock.ts", "**/*.test.{js,jsx,ts,tsx}"],
+  "ignoreUnresolved": [],
+  "ignoreUnimported": []
+}
