+++ libs/coin-modules/coin-stellar/CHANGELOG.md
@@ -1,0 +1,1 @@
+# @ledgerhq/coin-stellar
