+++ libs/coin-modules/coin-stellar/jest.config.js
@@ -1,0 +1,8 @@
+/** @type {import('ts-jest/dist/types').JestConfigWithTsJest} */
+module.exports = {
+  collectCoverageFrom: ["src/**/*.ts"],
+  coverageDirectory: "coverage",
+  preset: "ts-jest",
+  testEnvironment: "node",
+  testPathIgnorePatterns: ["lib/", "lib-es/"],
+};
