+++ libs/coin-modules/coin-stellar/src/errors.ts
@@ -1,8 +1,10 @@
 import { createCustomErrorClass } from "@ledgerhq/errors";
 
+export const StellarBurnAddressError = createCustomErrorClass("StellarBurnAddressError");
+export const StellarAssetRequired = createCustomErrorClass("StellarAssetRequired");
+export const StellarMuxedAccountNotExist = createCustomErrorClass("StellarMuxedAccountNotExist");
 export const StellarMemoRecommended = createCustomErrorClass("StellarMemoRecommended");
 export const StellarWrongMemoFormat = createCustomErrorClass("StellarWrongMemoFormat");
-export const StellarAssetRequired = createCustomErrorClass("StellarAssetRequired");
 export const StellarAssetNotAccepted = createCustomErrorClass("StellarAssetNotAccepted");
 export const StellarAssetNotFound = createCustomErrorClass("StellarAssetNotFound");
 export const StellarNotEnoughNativeBalance = createCustomErrorClass(
@@ -15,6 +17,4 @@ export const StellarFeeSmallerThanBase = createCustomErrorClass("StellarFeeSmall
 export const StellarNotEnoughNativeBalanceToAddTrustline = createCustomErrorClass(
   "StellarNotEnoughNativeBalanceToAddTrustline",
 );
-export const StellarMuxedAccountNotExist = createCustomErrorClass("StellarMuxedAccountNotExist");
 export const StellarSourceHasMultiSign = createCustomErrorClass("StellarSourceHasMultiSign");
-export const StellarBurnAddressError = createCustomErrorClass("StellarBurnAddressError");
