+++ libs/coin-modules/coin-stellar/src/estimateMaxSpendable.ts
@@ -3,8 +3,8 @@ import type { AccountBridge } from "@ledgerhq/types-live";
 import { getTransactionStatus } from "./getTransactionStatus";
 import { prepareTransaction } from "./prepareTransaction";
 import { createTransaction } from "./createTransaction";
-import { getMainAccount } from "../../account";
 import type { Transaction } from "./types";
+import { getMainAccount } from "@ledgerhq/coin-framework/account";
 
 const notCreatedStellarMockAddress = "GAW46JE3SHIAYLNNNQCAZFQ437WB5ZH7LDRDWR5LVDWHCTHCKYB6RCCH";
 
@@ -14,15 +14,16 @@ export const estimateMaxSpendable: AccountBridge<Transaction>["estimateMaxSpenda
   transaction,
 }) => {
   const mainAccount = getMainAccount(account, parentAccount);
-  const t = await prepareTransaction(mainAccount, {
+  const preparedTransaction = await prepareTransaction(mainAccount, {
     ...createTransaction(account),
     ...transaction,
     recipient: transaction?.recipient || notCreatedStellarMockAddress,
     // not used address
     useAllAmount: true,
   });
-  const s = await getTransactionStatus(mainAccount, t);
-  return s.amount.gte(0) ? s.amount : new BigNumber(0);
+
+  const status = await getTransactionStatus(mainAccount, preparedTransaction);
+  return status.amount.gte(0) ? status.amount : new BigNumber(0);
 };
 
 export default estimateMaxSpendable;
