+++ libs/coin-modules/coin-stellar/src/prepareTransaction.ts
