+++ libs/coin-modules/coin-stellar/src/tokens.ts
@@ -1,10 +1,11 @@
 import BigNumber from "bignumber.js";
 import { emptyHistoryCache } from "@ledgerhq/coin-framework/account/index";
-import { listTokensForCryptoCurrency, findTokenById, parseCurrencyUnit } from "../../currencies";
 import type { BalanceAsset, StellarOperation } from "./types";
-import { encodeOperationId } from "../../operation";
 import type { CryptoCurrency, TokenCurrency } from "@ledgerhq/types-cryptoassets";
 import type { SyncConfig, TokenAccount } from "@ledgerhq/types-live";
+import { parseCurrencyUnit } from "@ledgerhq/coin-framework/currencies/parseCurrencyUnit";
+import { encodeOperationId } from "@ledgerhq/coin-framework/operation";
+import { findTokenById, listTokensForCryptoCurrency } from "@ledgerhq/cryptoassets";
 
 export const getAssetIdFromTokenId = (tokenId: string): string => tokenId.split("/")[2];
