+++ libs/coin-modules/coin-stellar/src/types/signer.ts
@@ -1,0 +1,3 @@
+import type Stellar from "@ledgerhq/hw-app-str";
+
+export interface StellarSigner extends Stellar {}
