+++ libs/ledger-live-common/scripts/sync-families-dispatch.mjs
@@ -29,6 +29,7 @@ const familiesWPackage = [
   "near",
   "polkadot",
   "solana",
+  "stellar",
   "tezos",
   "tron",
   "xrp",
