+++ libs/ledger-live-common/src/errors.ts
@@ -125,7 +125,6 @@ export const BluetoothNotSupportedError = createCustomErrorClass("FwUpdateBlueto
 export const EConnResetError = createCustomErrorClass("EConnReset");
 
 export { ClaimRewardsFeesWarning } from "@ledgerhq/errors";
-export * from "./families/stellar/errors";
 export * from "@ledgerhq/coin-framework/errors";
 export * from "@ledgerhq/coin-algorand/errors";
 export * from "@ledgerhq/coin-bitcoin/errors";
@@ -134,4 +133,5 @@ export * from "@ledgerhq/coin-evm/errors";
 export * from "@ledgerhq/coin-near/errors";
 export * from "@ledgerhq/coin-polkadot/errors";
 export * from "@ledgerhq/coin-solana/errors";
+export * from "@ledgerhq/coin-stellar/errors";
 export * from "@ledgerhq/coin-tezos/errors";
